[Adjustment][Ezra] shadow for android
